requirements.txt
attrs==19.3.0
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-WTF==0.14.3
h5py==2.10.0
hdmf==1.6.4
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
numpy==1.19.0
pandas==1.0.5
pynwb==1.3.2
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.5.0
six==1.15.0
Werkzeug==1.0.1
WTForms==2.3.1
zipp==3.1.0
